+++ NEWS.rst
@@ -6,7 +6,8 @@ of every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Added support for the emscripten compiler.
+* Minor: Added emscripten mkspecs: cxx_emscripten126 and cxx_emscripten125.
 
 2.36.1
 ------